+++ src/stock-trading-vue/static/StockTradingApp.vue
@@ -1,8 +1,12 @@
 <script setup lang="ts">
+import DashboardView from './views/DashboardView.vue'
 import LoginView from './views/LoginView.vue'
 
+// TODO: authenticate user 
+const is_authenticated = true;
 </script>
 
 <template>
-    <LoginView />
+    <DashboardView v-if="is_authenticated" />
+    <LoginView v-else />
 </template>